Project/requirements.txt
numpy~=1.17.1
opencv-python~=4.1.1.26
Pillow~=6.1.0
matplotlib~=3.1.1